+++ src/layman/layer/geoserver/__init__.py
@@ -4,11 +4,11 @@ from flask import g
 import crs as crs_def
 from geoserver import util as gs_util
 from layman.http import LaymanError
-from layman import settings, util as layman_util
+from layman import settings, util as layman_util, names
 from layman.common import bbox as bbox_util, geoserver as gs_common
 from layman.layer import LAYER_TYPE
 from . import wms
-from .util import get_external_db_store_name
+from .util import get_external_db_store_name, get_internal_db_store_name
 
 logger = logging.getLogger(__name__)
 FLASK_RULES_KEY = f"{__name__}:RULES"
@@ -25,9 +25,9 @@ def delete_whole_user(username, auth=settings.LAYMAN_GS_AUTH):
 
 
 def ensure_workspace(workspace, auth=settings.LAYMAN_GS_AUTH):
-    created = gs_util.ensure_workspace(workspace, auth)
-    if created:
-        gs_util.create_db_store(workspace, auth, workspace, pg_conn=settings.PG_CONN)
+    db_store_name = get_internal_db_store_name(db_schema=workspace)
+    gs_util.ensure_db_store(names.GEOSERVER_WFS_WORKSPACE, auth, db_schema=workspace, pg_conn=settings.PG_CONN, name=db_store_name)
+    gs_util.ensure_db_store(names.GEOSERVER_WMS_WORKSPACE, auth, db_schema=workspace, pg_conn=settings.PG_CONN, name=db_store_name)
 
 
 def create_external_db_store(workspace, *, uuid, table_uri, auth=settings.LAYMAN_GS_AUTH):
@@ -41,7 +41,7 @@ def create_external_db_store(workspace, *, uuid, table_uri, auth=settings.LAYMAN
     store_name = get_external_db_store_name(uuid=uuid)
     gs_util.create_db_store(workspace,
                             auth,
-                            table_uri.schema,
+                            db_schema=table_uri.schema,
                             pg_conn=pg_conn,
                             name=store_name,
                             )
@@ -49,7 +49,10 @@ def create_external_db_store(workspace, *, uuid, table_uri, auth=settings.LAYMAN
 
 
 def delete_workspace(workspace, auth=settings.LAYMAN_GS_AUTH):
-    pass
+    db_store_name = get_internal_db_store_name(db_schema=workspace)
+    gs_util.delete_db_store(names.GEOSERVER_WFS_WORKSPACE, auth, store_name=db_store_name)
+    gs_util.delete_db_store(names.GEOSERVER_WMS_WORKSPACE, auth, store_name=db_store_name)
+
 
 def get_all_rules(auth):
     key = FLASK_RULES_KEY